update US01 UniqueId
